feat(chat): update tests and add new discord map
