+++ superface/super.json
@@ -585,6 +585,9 @@
         },
         "slack": {
           "file": "../grid/chat/delete-message/maps/slack.suma"
+        },
+        "discord": {
+          "file": "../grid/chat/delete-message/maps/discord.suma"
         }
       }
     },
